EDM-768: Only set pdeathsig on Linux
