+++ pkg/executer/executer.go
@@ -29,12 +29,6 @@ func (e *CommonExecuter) Execute(command string, args ...string) (stdout string,
 	return e.execute(context.Background(), cmd)
 }
 
-func (e *CommonExecuter) CommandContext(ctx context.Context, command string, args ...string) *exec.Cmd {
-	cmd := exec.CommandContext(ctx, command, args...)
-	cmd.SysProcAttr = &syscall.SysProcAttr{Pdeathsig: syscall.SIGTERM}
-	return cmd
-}
-
 func (e *CommonExecuter) ExecuteWithContext(ctx context.Context, command string, args ...string) (stdout string, stderr string, exitCode int) {
 	cmd := exec.CommandContext(ctx, command, args...)
 	return e.execute(ctx, cmd)
@@ -47,26 +41,6 @@ func (e *CommonExecuter) ExecuteWithContextFromDir(ctx context.Context, workingD
 		cmd.Env = env
 	}
 	return e.execute(ctx, cmd)
-}
-
-func (e *CommonExecuter) execute(ctx context.Context, cmd *exec.Cmd) (stdout string, stderr string, exitCode int) {
-	var stdoutBytes, stderrBytes bytes.Buffer
-	cmd.Stdout = &stdoutBytes
-	cmd.Stderr = &stderrBytes
-	// Set Pdeathsig to SIGTERM to kill the process and its children when the parent process is killed.
-	// This should prevent orphaned processes and allow for the subprocess to gracefully terminate.
-	// ref. https://github.com/golang/go/blob/release-branch.go1.21/src/syscall/exec_linux.go#L91
-	cmd.SysProcAttr = &syscall.SysProcAttr{Pdeathsig: syscall.SIGTERM}
-
-	if err := cmd.Run(); err != nil {
-		// handle timeout error
-		if errors.Is(ctx.Err(), context.DeadlineExceeded) {
-			return stdoutBytes.String(), context.DeadlineExceeded.Error(), 124
-		}
-		return stdoutBytes.String(), getErrorStr(err, &stderrBytes), getExitCode(err)
-	}
-
-	return stdoutBytes.String(), stderrBytes.String(), 0
 }
 
 func (e *CommonExecuter) LookPath(file string) (string, error) {
