+++ pkg/executer/executer_other.go
@@ -1,0 +1,31 @@
+//go:build !linux
+
+package executer
+
+import (
+	"bytes"
+	"context"
+	"errors"
+	"os/exec"
+)
+
+func (e *CommonExecuter) CommandContext(ctx context.Context, command string, args ...string) *exec.Cmd {
+	cmd := exec.CommandContext(ctx, command, args...)
+	return cmd
+}
+
+func (e *CommonExecuter) execute(ctx context.Context, cmd *exec.Cmd) (stdout string, stderr string, exitCode int) {
+	var stdoutBytes, stderrBytes bytes.Buffer
+	cmd.Stdout = &stdoutBytes
+	cmd.Stderr = &stderrBytes
+
+	if err := cmd.Run(); err != nil {
+		// handle timeout error
+		if errors.Is(ctx.Err(), context.DeadlineExceeded) {
+			return stdoutBytes.String(), context.DeadlineExceeded.Error(), 124
+		}
+		return stdoutBytes.String(), getErrorStr(err, &stderrBytes), getExitCode(err)
+	}
+
+	return stdoutBytes.String(), stderrBytes.String(), 0
+}
